+++ backend/canisters/local_user_index/api/src/updates/c2c_community_canister.rs
@@ -1,0 +1,2 @@
+pub type Args = crate::c2c_group_canister::Args;
+pub type Response = crate::c2c_group_canister::Response;