+++ backend/canisters/local_user_index/impl/src/updates/mod.rs
@@ -9,6 +9,7 @@ pub mod c2c_create_group;
 pub mod c2c_delete_community;
 pub mod c2c_delete_group;
 pub mod c2c_group_index;
+pub mod c2c_group_or_community_canister;
 pub mod c2c_local_group_index;
 pub mod c2c_notifications_index;
 pub mod c2c_notify_low_balance;